[#20] Transferred the initial countdown timer feature code to MainActivity.kt
